Start refactoring caches
