+++ pims/formats/common/sis.py
@@ -11,7 +11,7 @@
 #  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  * See the License for the specific language governing permissions and
 #  * limitations under the License.
-from pims.cache import cached_property
+from pims.cache.object import cached_property
 
 from pims.formats.utils.abstract import AbstractFormat, CachedDataPath
 from pims.formats.utils.engines.tifffile import TifffileChecker, TifffileParser, cached_tifffile
