+++ pims/formats/common/tiff.py
@@ -15,7 +15,7 @@ from typing import List, Optional, Union
 
 from pyvips import Image as VIPSImage
 
-from pims.cache import cached_property
+from pims.cache.object import cached_property
 from pims.formats import AbstractFormat
 from pims.formats.utils.abstract import CachedDataPath
 from pims.formats.utils.engines.tifffile import TIFF_FLAGS, TifffileChecker, TifffileParser
