+++ pims/formats/common/webp.py
@@ -16,7 +16,7 @@ from typing import Optional
 
 from pint import Quantity
 
-from pims.cache import cached_property
+from pims.cache.object import cached_property
 from pims.formats import AbstractFormat
 from pims.formats.utils.abstract import CachedDataPath
 from pims.formats.utils.checker import SignatureChecker
